Refactor iOS Auth layer to use real KMP use cases
- Consolidated parallel AuthViewModel structures into `Features/Auth/ViewModels/AuthViewModel.swift`.
- Added `signInUseCase`, `signUpUseCase`, and `sendPasswordResetUseCase` initialization in `DependencyContainer`.
- Wired KMP shared UseCases for actual networking instead of `DispatchQueue.asyncAfter` mock delays.
- Removed "Simulate" labels from `LoginView` and `RegisterView` and added accurate text fields.
- Removed duplicated `AuthViewModel` in the legacy Auth directory.
