+++ iosApp/iosApp/Core/DI/DependencyContainer.swift
@@ -8,14 +8,34 @@ class DependencyContainer: ObservableObject {
     let analyticsService: AnalyticsService
     let crashReportingService: CrashReportingService
 
+    // Auth Repos & Use Cases
+    let authRepository: AuthRepository
+    let signInUseCase: SignInUseCase
+    let signUpUseCase: SignUpUseCase
+    let sendPasswordResetUseCase: SendPasswordResetUseCase
+
     private init() {
         self.analyticsService = DefaultAnalyticsService()
         self.crashReportingService = DefaultCrashReportingService()
+
+        // Initialize KMP Shared Auth Use Cases
+        let client = SupabaseClient.shared.client
+        self.authRepository = SupabaseAuthRepository(client: client)
+        self.signInUseCase = SignInUseCase(repository: self.authRepository)
+        self.signUpUseCase = SignUpUseCase(repository: self.authRepository)
+        self.sendPasswordResetUseCase = SendPasswordResetUseCase(repository: self.authRepository)
     }
 
     // ViewModels
+    @MainActor
     func makeAuthViewModel() -> AuthViewModel {
-        return AuthViewModel(analyticsService: analyticsService)
+        let viewModel = AuthViewModel(analyticsService: analyticsService)
+        viewModel.setDependencies(
+            signInUseCase: signInUseCase,
+            signUpUseCase: signUpUseCase,
+            sendPasswordResetUseCase: sendPasswordResetUseCase
+        )
+        return viewModel
     }
 
     func makeHomeViewModel() -> HomeViewModel {
